Move all configuration into global config file
This helps with clarifying what is a config value, which config values
need to change, and makes it easier to test. These changes have been
verified by running a test set of results through the validator before
and after the change, and checking for no diff with the same config.
Now for the 2025 election most, if not all, of the changes can be made
just within this new config file.